# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:1400",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/portfolio/sent-mail": {
      "post": {
        "tags": [
          "My Portfolio"
        ],
        "description": "Endpoint to sent mail from my contact form on the porfolio",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Sent email.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SentMail"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/portfolio/mails": {
      "get": {
        "tags": [
          "My Portfolio"
        ],
        "description": "Endpoint to fetch all mails from my contact form on the porfolio",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "SentMail": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John"
        },
        "email": {
          "type": "string",
          "example": "email@example.com"
        },
        "message": {
          "type": "string",
          "example": "What you want to enquire"
        },
        "subject": {
          "type": "string",
          "example": "Enquiry"
        }
      },
      "required": [
        "email",
        "message",
        "name",
        "subject"
      ]
    },
...
  }
}